feat: add input for create new playlist
